Finished moving components
