+++ apps/frontend-react/src/components/VolumeSlider.module.css
