+++ apps/frontend-react/src/components/VolumeSlider.tsx
