+++ apps/frontend-react/src/components/stream/StreamImage.tsx
